# perf: use uint `MAX_INDEX` on `arm` / `aarch64` for uint SIMD