--- a/src/simd/simd_u64.rs
+++ b/src/simd/simd_u64.rs
@@ -51,12 +51,10 @@ fn _i64ord_to_u64(ord_i64: i64) -> u64 {
     unsafe { std::mem::transmute::<i64, u64>(ord_i64 ^ XOR_VALUE) }
 }
 
-#[cfg(any(
-    target_arch = "x86",
-    target_arch = "x86_64",
-    all(target_arch = "aarch64", feature = "nightly_simd")
-))]
-const MAX_INDEX: usize = i64::MAX as usize;
+#[cfg(any(target_arch = "x86", target_arch = "x86_64",))]
+const MAX_INDEX: usize = i64::MAX as usize; // SIMD operations on signed ints
+#[cfg(all(target_arch = "aarch64", feature = "nightly_simd"))]
+const MAX_INDEX: usize = u64::MAX as usize; // SIMD operations on unsigned ints
 
 // --------------------------------------- AVX2 ----------------------------------------